Add new structure and docs
I think it would be really helpful to have a brief overview of all
of the different projects that are going on, where they're hosted
and how they interact with each other.

This will make onboarding much easier and will hopefully give everyone
a much better idea of what's going on.