Move platform doc into central "tech" section
I've tried to remove the bits that I don't think are relevant here:
there's no mention of building containers for Django apps which I
think is quite important to how we work now.

However I've left some things in which I'm not sure about, so there
might still be some things in this commit which aren't currently
accurate.